Remove old diskmanager
