+++ Tests/Kernels/Cosmos.Kernel.Tests.DiskManager/Kernel.cs
@@ -12,7 +12,7 @@ namespace Cosmos.Kernel.Tests.DiskManager
     /// <seealso cref="Cosmos.System.Kernel" />
     public class Kernel : Sys.Kernel
     {
-        private VFSBase mVFS;
+        public static CosmosVFS mVFS;
 
         /// <summary>
         /// Pre-run events
